Restructre managers to clipper class
