--- packages/geometry-utils/src/clipper/out-rec-manager.ts
@@ -1,387 +1,0 @@
-import { join_u16_to_u32, get_u16_from_u32 } from 'wasm-nesting';
-import { Point } from '../types';
-import Join from './join';
-import { DIRECTION } from './types';
-import { PointI32 } from '../geometry';
-import { UNASSIGNED } from './constants';
-import TEdge from './t-edge';
-import OutRec from './out-rec';
-
-export default class OutRecManager {
-    private join: Join;
-    private tEdge: TEdge;
-    private outRec: OutRec;
-
-    constructor(tEdgeController: TEdge, reverseSolution: boolean, strictlySimple: boolean) {
-        this.join = new Join();
-        this.tEdge = tEdgeController;
-        this.outRec = new OutRec(reverseSolution, strictlySimple);
-    }
-
-    public get strictlySimple(): boolean {
-        return this.outRec.strictlySimple;
-    }
-
-    public insertJoin(
-        condition: boolean,
-        outHash1: number,
-        edgeIndex: number,
-        point1: Point<Int32Array>,
-        point2: Point<Int32Array> = point1
-    ): boolean {
-        if (condition) {
-            const outHash2 = this.addOutPt(edgeIndex, point1);
-            this.join.add(outHash1, outHash2, point2);
-        }
-
-        return condition;
-    }
-
-    public addScanbeamJoin(edge1Index: number, edge2Index: number, point: Point<Int32Array>): void {
-        const outPt1 = this.addOutPt(edge2Index, point);
-        const outPt2 = this.addOutPt(edge1Index, point);
-
-        this.join.add(outPt1, outPt2, point);
-        //StrictlySimple (type-3) join
-    }
-
-    public addOutputJoins(outHash: number, rightBoundIndex: number): void {
-        const joinCount: number = this.join.getLength(true);
-
-        if (joinCount > 0) {
-            const point = PointI32.create();
-
-            for (let i = 0; i < joinCount; ++i) {
-                //if the horizontal Rb and a 'ghost' horizontal overlap, then convert
-                //the 'ghost' join to a real join ready for later ...
-                point.set(this.join.getX(i, true), this.join.getY(i, true));
-
-                if (this.horzSegmentsOverlap(this.join.getHash1(i, true), point, rightBoundIndex)) {
-                    this.join.fromGhost(i, outHash);
-                }
-            }
-        }
-    }
-
-    public reset() {
-        this.join.reset();
-    }
-
-    public prepareHorzJoins(horzEdgeIndex: number) {
-        //Also, since horizontal edges at the top of one SB are often removed from
-        //the AEL before we process the horizontal edges at the bottom of the next,
-        //we need to create 'ghost' Join records of 'contrubuting' horizontals that
-        //we can compare with horizontals at the bottom of the next SB.
-        //get the last Op for this horizontal edge
-        //the point may be anywhere along the horizontal ...
-        const [outPtHash, x, y] = this.getJoinData(horzEdgeIndex);
-
-        this.join.addGhost(outPtHash, x, y);
-    }
-
-    public clearGhostJoins() {
-        this.join.clearGhosts();
-    }
-
-    public addOutPt(edgeIndex: number, point: Point<Int32Array>): number {
-        let outRecIndex: number;
-        let pointIndex: number;
-
-        if (!this.tEdge.isAssigned(edgeIndex)) {
-            pointIndex = this.outRec.fromPoint(point);
-
-            outRecIndex = this.outRec.create(pointIndex);
-
-            this.setHoleState(outRecIndex, edgeIndex);
-
-            this.tEdge.setRecIndex(edgeIndex, this.outRec.currentIndex(outRecIndex));
-            //nb: do this after SetZ !
-        } else {
-            const isToFront: boolean = this.tEdge.side(edgeIndex) === DIRECTION.LEFT;
-            const recIndex = this.tEdge.getRecIndex(edgeIndex);
-            outRecIndex = this.outRec.getOutRec(recIndex);
-
-            pointIndex = this.outRec.addOutPt(recIndex, isToFront, point);
-        }
-
-        return this.outRec.getHash(outRecIndex, pointIndex);
-    }
-
-    public addLocalMinPoly(edge1Index: number, edge2Index: number, point: Point<Int32Array>): number {
-        let firstIndex = edge2Index;
-        let secondIndex = edge1Index;
-        let result: number = UNASSIGNED;
-
-        if (this.tEdge.isHorizontal(edge2Index) || this.tEdge.dx(edge1Index) > this.tEdge.dx(edge2Index)) {
-            firstIndex = edge1Index;
-            secondIndex = edge2Index;
-        }
-
-        result = this.addOutPt(firstIndex, point);
-        this.tEdge.setRecIndex(secondIndex, this.tEdge.getRecIndex(firstIndex));
-        this.tEdge.setSide(secondIndex, DIRECTION.RIGHT);
-        this.tEdge.setSide(firstIndex, DIRECTION.LEFT);
-
-        const prevIndex =
-            this.tEdge.prevActive(firstIndex) === secondIndex
-                ? this.tEdge.prevActive(secondIndex)
-                : this.tEdge.prevActive(firstIndex);
-        const condition = this.tEdge.checkMinJoin(firstIndex, prevIndex, point);
-
-        this.insertJoin(condition, result, prevIndex, point, this.tEdge.top(firstIndex));
-
-        return result;
-    }
-
-    public addLocalMaxPoly(edge1Index: number, edge2Index: number, point: Point<Int32Array>): void {
-        this.addOutPt(edge1Index, point);
-
-        if (this.tEdge.isWindDeletaEmpty(edge2Index)) {
-            this.addOutPt(edge2Index, point);
-        }
-
-        const recIndex1 = this.tEdge.getRecIndex(edge1Index);
-        const recIndex2 = this.tEdge.getRecIndex(edge2Index);
-
-        if (recIndex1 === recIndex2) {
-            this.tEdge.unassign(edge1Index);
-            this.tEdge.unassign(edge2Index);
-            return;
-        }
-
-        const condition = recIndex1 < recIndex2;
-        const firstIndex = condition ? edge1Index : edge2Index;
-        const secondIndex = condition ? edge2Index : edge1Index;
-        const firstRecIndex = this.tEdge.getRecIndex(firstIndex);
-        const secondRecIndex = this.tEdge.getRecIndex(secondIndex);
-
-        //get the start and ends of both output polygons ...
-        const firstSide = this.tEdge.side(firstIndex);
-        const secondSide = this.tEdge.side(secondIndex);
-
-        this.outRec.joinPolys(firstRecIndex, secondRecIndex, firstSide, secondSide);
-        
-        const OKIdx: number = this.tEdge.getRecIndex(firstIndex);
-        const ObsoleteIdx: number = this.tEdge.getRecIndex(secondIndex);
-        this.tEdge.unassign(firstIndex);
-        //nb: safe because we only get here via AddLocalMaxPoly
-        this.tEdge.unassign(secondIndex);
-
-        this.tEdge.updateIndexAEL(firstSide, ObsoleteIdx, OKIdx);
-
-        this.outRec.setCurrentIndex(secondRecIndex, firstRecIndex);
-    }
-
-    public fixupOutPolygon(): void {
-        let i: number = 0;
-
-        this.outRec.fixDirections();
-
-        const joinCount: number = this.join.getLength(false);
-        const point = PointI32.create();
-
-        for (i = 0; i < joinCount; ++i) {
-            point.set(this.join.getX(i, false), this.join.getY(i, false));
-            this.joinCommonEdge(i, point);
-        }
-
-        this.outRec.fixOutPolygon(this.tEdge.isUseFullRange);
-    }
-
-    public buildResult(polygons: Point<Int32Array>[][]): void {
-        return this.outRec.buildResult(polygons);
-    }
-
-    public dispose(): void {
-        this.outRec.dispose();
-    }
-
-    private getJoinData(index: number) {
-        //get the last Op for this horizontal edge
-        //the point may be anywhere along the horizontal ...
-        const recIndex = this.tEdge.getRecIndex(index);
-        const side = this.tEdge.side(index);
-        const top = this.tEdge.top(index);
-        const bot = this.tEdge.bot(index);
-
-        return this.outRec.getJoinData(recIndex, side, top, bot);
-    }
-
-    private setHoleState(recIndex: number, edgeIndex: number): void {
-        const { isHole, index } = this.tEdge.getHoleState(this.outRec.firstLeftIndex(recIndex), edgeIndex);
-
-        this.outRec.setHoleState(recIndex, isHole, index);
-    }
-
-    private joinCommonEdge(index: number, offPoint: Point<Int32Array>): void {
-        const inputHash1 = this.join.getHash1(index, false);
-        const inputHash2 = this.join.getHash2(index);
-        const { outHash1, outHash2, result } = this.joinPoints(inputHash1, inputHash2, offPoint);
-
-        if (!result) {
-            this.join.updateHash(index, outHash1, outHash2);
-            return;
-        }
-
-        //get the polygon fragment with the correct hole state (FirstLeft)
-        //before calling JoinPoints() ...
-
-        const index1: number = get_u16_from_u32(outHash1, 0);
-        const index2: number = get_u16_from_u32(outHash2, 0);
-        const outPt1Index: number = get_u16_from_u32(outHash1, 1);
-        const outPt2Index: number = get_u16_from_u32(outHash2, 1);
-        const outRec1: number = this.outRec.getOutRec(index1);
-        let outRec2: number = this.outRec.getOutRec(index2);
-
-        if (index1 === index2) {
-            //instead of joining two polygons, we've just created a new one by
-            //splitting one polygon into two.
-            outRec2 = this.outRec.splitPolys(outRec1, outPt1Index, outPt2Index);
-
-            this.join.updateHash(index, outHash1, outHash2);
-            return;
-        }
-
-        this.outRec.joinPolys2(outRec1, outRec2);
-
-        this.join.updateHash(index, outHash1, outHash2);
-    }
-
-    private joinPoints(
-        outHash1: number,
-        outHash2: number,
-        offPoint: Point<Int32Array>
-    ): { outHash1: number; outHash2: number; result: boolean } {
-        const index1: number = get_u16_from_u32(outHash1, 0);
-        const index2: number = get_u16_from_u32(outHash2, 0);
-        const outRec1 = this.outRec.getOutRec(index1);
-        const outRec2 = this.outRec.getOutRec(index2);
-        const result = { outHash1, outHash2, result: false };
-
-        if (this.outRec.isUnassigned(outRec1) || this.outRec.isUnassigned(outRec2)) {
-            return result;
-        }
-
-        const outPt1Index: number = get_u16_from_u32(outHash1, 1);
-        const outPt2Index: number = get_u16_from_u32(outHash2, 1);
-        const isRecordsSame = outRec1 === outRec2;
-        //There are 3 kinds of joins for output polygons ...
-        //1. Horizontal joins where Join.OutPt1 & Join.OutPt2 are a vertices anywhere
-        //along (horizontal) collinear edges (& Join.OffPt is on the same horizontal).
-        //2. Non-horizontal joins where Join.OutPt1 & Join.OutPt2 are at the same
-        //location at the Bottom of the overlapping segment (& Join.OffPt is above).
-        //3. StrictlySimple joins where edges touch but are not collinear and where
-        //Join.OutPt1, Join.OutPt2 & Join.OffPt all share the same point.
-        const isHorizontal: boolean = this.outRec.point(outPt1Index).y === offPoint.y;
-
-        if (
-            isHorizontal &&
-            offPoint.almostEqual(this.outRec.point(outPt1Index)) &&
-            offPoint.almostEqual(this.outRec.point(outPt2Index))
-        ) {
-            //Strictly Simple join ...
-            const reverse1 = this.outRec.strictlySimpleJoin(outPt1Index, offPoint);
-            const reverse2 = this.outRec.strictlySimpleJoin(outPt2Index, offPoint);
-
-            if (reverse1 === reverse2) {
-                return result;
-            }
-
-            result.outHash2 = join_u16_to_u32(index2, this.outRec.applyJoin(outPt1Index, outPt2Index, reverse1));
-            result.result = true;
-
-            return result;
-        }
-
-        if (isHorizontal) {
-            //treat horizontal joins differently to non-horizontal joins since with
-            //them we're not yet sure where the overlapping is. OutPt1.Pt & OutPt2.Pt
-            //may be anywhere along the horizontal edge.
-            const outPt1Res = this.outRec.flatHorizontal(outPt1Index, outPt2Index, outPt2Index);
-
-            if (outPt1Res.length === 0) {
-                return result;
-            }
-
-            const [op1Index, op1bIndex] = outPt1Res;
-            //a flat 'polygon'
-            const outPt2Res = this.outRec.flatHorizontal(outPt2Index, op1Index, op1bIndex);
-
-            if (outPt2Res.length === 0) {
-                return result;
-            }
-
-            const [op2Index, op2bIndex] = outPt2Res;
-            //a flat 'polygon'
-            //Op1 -. Op1b & Op2 -. Op2b are the extremites of the horizontal edges
-
-            const value = PointI32.getOverlap(
-                this.outRec.pointX(op1Index),
-                this.outRec.pointX(op1bIndex),
-                this.outRec.pointX(op2Index),
-                this.outRec.pointX(op2bIndex)
-            );
-            const isOverlapped = value.x < value.y;
-
-            if (!isOverlapped) {
-                return result;
-            }
-
-            //DiscardLeftSide: when overlapping edges are joined, a spike will created
-            //which needs to be cleaned up. However, we don't want Op1 or Op2 caught up
-            //on the discard Side as either may still be needed for other joins ...
-            result.outHash1 = join_u16_to_u32(index1, op1Index);
-            result.outHash2 = join_u16_to_u32(index2, op2Index);
-            result.result = this.outRec.joinHorz(op1Index, op1bIndex, op2Index, op2bIndex, value);
-
-            return result;
-        }
-
-        let op1 = outPt1Index;
-        let op2 = outPt2Index;
-        let op1b: number = this.outRec.getUniquePt(op1, true);
-        let op2b: number = this.outRec.getUniquePt(op2, true);
-        //nb: For non-horizontal joins ...
-        //    1. Jr.OutPt1.Pt.Y === Jr.OutPt2.Pt.Y
-        //    2. Jr.OutPt1.Pt > Jr.OffPt.Y
-        //make sure the polygons are correctly oriented ...
-
-        const reverse1: boolean = this.tEdge.checkReverse(this.outRec.point(op1), this.outRec.point(op1b), offPoint);
-
-        if (reverse1) {
-            op1b = this.outRec.getUniquePt(op1, false);
-
-            if (this.tEdge.checkReverse(this.outRec.point(op1), this.outRec.point(op1b), offPoint)) {
-                return result;
-            }
-        }
-
-        const reverse2: boolean = this.tEdge.checkReverse(this.outRec.point(op2), this.outRec.point(op2b), offPoint);
-
-        if (reverse2) {
-            op2b = this.outRec.getUniquePt(op2, false);
-
-            if (this.tEdge.checkReverse(this.outRec.point(op2), this.outRec.point(op2b), offPoint)) {
-                return result;
-            }
-        }
-
-        if (op1b === op1 || op2b === op2 || op1b === op2b || (isRecordsSame && reverse1 === reverse2)) {
-            return result;
-        }
-
-        result.outHash2 = join_u16_to_u32(index2, this.outRec.applyJoin(outPt1Index, outPt2Index, reverse1));
-
-        result.result = true;
-
-        return result;
-    }
-
-    private horzSegmentsOverlap(outHash: number, offPoint: Point<Int32Array>, edgeIndex: number): boolean {
-        const outPtIndex = get_u16_from_u32(outHash, 1);
-        const top = this.tEdge.top(edgeIndex);
-        const bot = this.tEdge.bot(edgeIndex);
-
-        return PointI32.horzSegmentsOverlap(this.outRec.point(outPtIndex), offPoint, bot, top);
-    }
-}
